[FIX] 코드 개편
