+++ index.html
@@ -2,27 +2,13 @@
 <html lang="ko">
   <head>
     <meta charset="UTF-8">
-    <link rel="icon" href="/favicon.ico">
+    <link rel="icon" href="/favicon.png">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>test</title>
+    <title>단어 말하기</title>
     <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
   </head>
   <body>
-    <!-- <video id="videoInput" style="display:none"></video>
-    <canvas id="videoOutput" style="width:500px; height: 280;"></canvas>
-    <button onclick=stream()>Send</button> -->
-    <div id="app"></div>
+    <div id="app" style="min-width: 90vw;"></div>
     <script type="module" src="/src/main.js"></script>
-    <!-- <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700" type="text/css">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons"> -->
-    <!-- <link rel="stylesheet" href="https://code.getmdl.io/1.3.0/material.indigo-pink.min.css">  -->
-        
-    <!-- Compiled and minified CSS -->
-    <!-- <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css"> -->
-
-    <!-- Compiled and minified JavaScript -->
-    <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script> -->
-        
-    <!-- <link href="{{ url_for('static', path='/styles.css') }}" rel="stylesheet"> -->
   </body>
 </html>
