+++ src/assets/gamecomp/minBtn.png
