--- src/views/HomeView.vue
@@ -1,9 +1,0 @@
-<script setup>
-import TheWelcome from "../components/TheWelcome.vue";
-</script>
-
-<template>
-    <main>
-        <TheWelcome />
-    </main>
-</template>
