+++ vite.config.js
@@ -15,7 +15,8 @@ export default defineConfig({
     },
     server: {
         proxy: {
-            "/api": "https://webdev-test.site/",
+            // "/api": "https://webdev-test.site/",
+            "/api": "localhost:8000/",
         },
     },
 });
